remove path params aliasing
