+++ code/frameworks/nextjs/package.json
@@ -32,16 +32,6 @@
       "require": "./dist/image-context.js",
       "import": "./dist/image-context.mjs"
     },
-    "./dist/routing/path-params-provider": {
-      "types": "./dist/routing/path-params-provider.d.ts",
-      "require": "./dist/routing/path-params-provider.js",
-      "import": "./dist/routing/path-params-provider.mjs"
-    },
-    "./dist/routing/path-params-provider-mock": {
-      "types": "./dist/routing/path-params-provider-mock.d.ts",
-      "require": "./dist/routing/path-params-provider-mock.js",
-      "import": "./dist/routing/path-params-provider-mock.mjs"
-    },
     "./preset": {
       "types": "./dist/preset.d.ts",
       "require": "./dist/preset.js"
@@ -69,9 +59,6 @@
       ],
       "dist/image-context": [
         "dist/image-context.d.ts"
-      ],
-      "dist/routing/path-params-provider": [
-        "dist/routing/path-params-provider-mock.d.ts"
       ]
     }
   },
@@ -177,8 +164,6 @@
       "./src/images/decorator.tsx",
       "./src/images/next-legacy-image.tsx",
       "./src/images/next-image.tsx",
-      "./src/routing/path-params-provider.tsx",
-      "./src/routing/path-params-provider-mock.tsx",
       "./src/font/webpack/loader/storybook-nextjs-font-loader.ts",
       "./src/rsc/server-only.ts",
       "./src/swc/next-swc-loader-patch.ts"
